generalize extra models
